Test : Set the user data as mock to avoid type error
I saw this error -
TypeError: Cannot destructure property 'userName' of '(intermediate value)' as it is undefined.

it seemed to occur because data from db can be empty. but I won't manage the case now so that I can focus on other works. it is less necessary than working a lot of other things. to treat this case I won't let the data empty in DB.