Test&Comment : Test when there is any data except for empty case
I recognized that the ts error I wrote in last commit was occurred also because of mismatch of data in DB as I set them temporarily. so I edited the comment. anyway I'm going to treat the case where at least one data exist in order to save time and work on essential things