Fix : Send root comments without reComments
I'm changing how to treat comments.
at first I was trying to send comments divided with its page for when the talk content has comments too much,
but i was in stuck. I couldn't get the exact number of comments per comment page because reComments that have their parent comments couldn't be treated and cut from DB directly. in my previous code, I could get comments that were mixed between root comments and reComments and the reComments was composed later.

to solve that, I am trying to get comments in two steps.
first I send root comments without their reComments when a user visits talk detail page.
(root comments don't have parent comments for reComment)
in this time, a root comment has a property "reCommentNo" that is the number of reComments the root has.
in client side code, users will be able to see the number of reComments in one root comment (from property "reCommentNo")  and click the button to see reComments.
then, I will send reComments in the root comment from server side. this is the second step and I'm on it