Feat | Fix : Create an event to create and join a new room with partner
An api to get or create a roomId was replaced with this socket event
 though this event doesn't send roomId only.
Of course some of logic changed and will change in frontend work too.

An user will send this event when he/she clicks a chat icon in partner's page
 but the room is not in the chatSlice in the moment.
Then he/she will create and join the room.
 The partner will join too if he/she logged in too.
The user who send this event will receive and set the room in chatSlice with an empty message data as composing it.
The partner won't receive but will set it when a new message comes in the room that doesn't exist in the partner's chatSlice.
 (These are the frontend work)

And I removed previous routes and controllers for chat apis.